requirements.txt
PyQt6
Pillow
imagehash
opencv-python
numpy
pillow-heif
send2trash
piexif
torch
torchvision
